新增权限信息服务类
